requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.2
wikipedia>=1.4.0
nltk>=3.8.1
markdown>=3.4.0
pygments>=2.15.0
tiktoken>=0.4.0
openai>=0.27.0
pandas>=2.0.0
plotly>=5.14.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
yfinance>=0.2.18
wolframalpha>=5.0.0
spacy>=3.5.0
textblob>=0.17.1
networkx>=3.1
matplotlib>=3.7.0
wordcloud>=1.9.0
seaborn>=0.12.0
scikit-learn>=1.2.0 